Commit 24/09 - 18:35

+++ Assets/Scripts/Game/Dialogues/DialogueSO.cs
@@ -7,7 +7,7 @@ public class AwnserConsequence
     public List<string> DialogueAfterAwnser;
     public List<LegendStats> ConsequenceStats = new List<LegendStats>();
     public List<DialogueSO> ConsequenceDialogues = new List<DialogueSO>();
-    public float SoulsAmount;
+    public long SoulsAmount;
 
     [Header("Effector")]
     public List<GameObject> ConsequenceEffector;
@@ -19,7 +19,7 @@ public class DialogueSO : ScriptableObject
     public CharacterSO characterOwner;
 
     [Header("Souls")]
-    public float FreeSouls_END;
+    public long FreeSouls_END;
     public bool isTaxes;
 
     [Header("Dialogue")]
